made fixes in router packages

+++ src/business/router.py
@@ -1,13 +1,12 @@
 import os
 import shutil
-
 from fastapi import APIRouter, HTTPException, UploadFile, File, Depends
 import secrets
 from PIL import Image
 from tortoise.exceptions import DoesNotExist
 from src.business.schemas import user_business_pydantic, user_business_pydanticIn, user_product_pydantic,\
     user_product_pydanticIn, UserProduct, UserBusiness
-from src.database import db
+from src.user.schemas import User
 from datetime import datetime
 from src.user.services import get_current_user
 from src.exceptions import ServerErrorException, UserNotFoundException, ProductNotFoundException, UnauthorizedUserException,\
@@ -25,48 +24,56 @@ router = APIRouter(
 async def add_user_business(business_info: user_business_pydanticIn,
                             current_user=Depends(get_current_user)):
     try:
-        user = await db["users"].find_one({"_id": current_user["_id"]})
+        user = await User.get_or_none(id=current_user.id)
+        if user is None:
+            raise UserNotFoundException("User not found")
         business_details = business_info.dict(exclude_unset=True)
         business_details["creation_date"] = datetime.utcnow()
         business_details["last_updated"] = datetime.utcnow()
-        business_details["owner"] = user["_id"]
+        business_details["owner_id"] = user.id
         business_obj = await UserBusiness.create(**business_details)
         response = await user_business_pydantic.from_tortoise_orm(business_obj)
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
         raise UserNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
 
 
 @router.get('/all')
 async def get_user_businesses(current_user=Depends(get_current_user)):
     try:
-        response = await user_business_pydantic.from_queryset(UserBusiness.filter(owner=current_user.id))
+        response = await user_business_pydantic.from_queryset(UserBusiness.filter(owner=current_user))
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        BusinessNotFoundException(str(dne))
+        raise BusinessNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
 
 
 @router.get('/get/{business_id}')
-async def get_specific_user_business(business_id: int):
+async def get_specific_user_business(business_id: int, current_user=Depends(get_current_user)):
     if business_id == 0:
         raise InvalidIdException("Invalid business id")
     try:
+        business = await UserBusiness.get(id=business_id)
+        if business.owner_id != current_user.id:
+            raise UnauthorizedUserException("User not authorized")
         response = await user_business_pydantic.from_queryset_single(UserBusiness.get(id=business_id))
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        BusinessNotFoundException(str(dne))
+        raise BusinessNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
 
 
 @router.put('/update/{business_id}')
-async def update_user_business(business_id: int, update_info: user_business_pydanticIn):
+async def update_user_business(business_id: int, update_info: user_business_pydanticIn,
+                               current_user=Depends(get_current_user)):
     try:
         business = await UserBusiness.get(id=business_id)
+        if business.owner_id != current_user.id:
+            raise UnauthorizedUserException("User not authorized")
         update_info = update_info.dict(exclude_unset=True)
         business.business_name = update_info['business_name']
         business.logo = update_info['logo']
@@ -79,9 +86,10 @@ async def update_user_business(business_id: int, update_info: user_business_pyda
         response = await user_business_pydantic.from_tortoise_orm(business)
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        BusinessNotFoundException(str(dne))
+        raise BusinessNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
+
 
 
 @router.delete('/delete/{business_id}')
@@ -89,7 +97,8 @@ async def delete_user_business(business_id: int, current_user=Depends(get_curren
     if business_id == 0:
         raise InvalidIdException("Invalid business id")
     business = await UserBusiness.get(id=business_id)
-    if current_user["username"] not in ["root", "master"] or current_user.id != business.owner:
+    owner = await User.filter(id=business.owner_id).first()
+    if current_user.username not in ["root", "master"] or current_user.id != owner.id:
         raise HTTPException(status_code=403, detail="User not authorized")
     try:
         try:
@@ -110,49 +119,59 @@ async def delete_user_business(business_id: int, current_user=Depends(get_curren
         ServerErrorException(str(e))
 
 
+
+
 @router.post('/product/add/{business_id}')
-async def add_user_product(business_id: int, product_details: user_product_pydanticIn):
+async def add_user_product(business_id: int, product_details: user_product_pydanticIn, current_user=Depends(get_current_user)):
     try:
         business = await UserBusiness.get(id=business_id)
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
         product_details = product_details.dict(exclude_unset=True)
         business_obj = await UserProduct.create(**product_details, business=business)
         response = await user_product_pydantic.from_tortoise_orm(business_obj)
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        BusinessNotFoundException(str(dne))
+        raise BusinessNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
-
+        raise ServerErrorException(str(e))
 
 @router.get('/{business_id}/get/products')
-async def get_all_business_products(business_id: int):
+async def get_all_business_products(business_id: int, current_user=Depends(get_current_user)):
     if business_id == 0:
         raise InvalidIdException("Invalid product id")
     try:
         business = await UserBusiness.get(id=business_id)
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
         response = await user_product_pydantic.from_queryset(UserProduct.filter(business=business))
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        BusinessNotFoundException(str(dne))
+        raise BusinessNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
-
+        raise ServerErrorException(str(e))
 
 @router.get('/product/get/{product_id}')
-async def get_specific_user_product(product_id: int):
+async def get_specific_user_product(product_id: int, current_user=Depends(get_current_user)):
     try:
+        product = await UserProduct.get(id=product_id)
+        business = await product.business
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
         response = await user_product_pydantic.from_queryset_single(UserProduct.get(id=product_id))
         return {"status": "ok", "data": response}
     except DoesNotExist as dne:
-        ProductNotFoundException(str(dne))
+        raise ProductNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
-
+        raise ServerErrorException(str(e))
 
 @router.put('/product/update/{product_id}')
-async def update_user_product(product_id: int, update_info: user_product_pydanticIn):
+async def update_user_product(product_id: int, update_info: user_product_pydanticIn, current_user=Depends(get_current_user)):
     try:
         product = await UserProduct.get(id=product_id)
+        business = await product.business
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
         update_info = update_info.dict(exclude_unset=True)
         product.name = update_info['name']
         product.category = update_info['category']
@@ -170,31 +189,43 @@ async def update_user_product(product_id: int, update_info: user_product_pydanti
 
 
 @router.delete('/product/delete/{product_id}')
-async def delete_user_product(product_id: int):
+async def delete_user_product(product_id: int, current_user=Depends(get_current_user)):
     if product_id == 0:
         raise InvalidIdException("Invalid product id")
     try:
+        product = await UserProduct.get(id=product_id)
+        business = await product.business
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
         await UserProduct.filter(id=product_id).delete()
-        static_folder = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))),
-                                     'static')
+
+        # Get the base path of your project
+        base_folder = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+        static_folder = os.path.join(base_folder, 'static')
         folder_path = os.path.join(static_folder, f"user_products/{product_id}")
         if os.path.exists(folder_path):
             shutil.rmtree(folder_path)
         return {"status": "ok", "message": "product successfully deleted"}
     except DoesNotExist as dne:
-        ProductNotFoundException(str(dne))
+        raise ProductNotFoundException(str(dne))
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
+
 
 
 @router.post("/product/upload/{product_id}")
-async def create_upload_file(product_id: int, file: UploadFile = File(...),
-                             user=Depends(get_current_user)):
+async def create_upload_file(product_id: int, file: UploadFile = File(...), current_user=Depends(get_current_user)):
     if product_id == 0:
         raise InvalidIdException("Invalid product id")
     try:
-        static_folder = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))),
-                                     'static')
+        product = await UserProduct.get(id=product_id)
+        business = await product.business
+        if current_user.id != business.owner_id:
+            raise UnauthorizedUserException("User not authorized")
+
+        # Get the base path of your project
+        base_folder = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+        static_folder = os.path.join(base_folder, 'static')
         filepath = os.path.join(static_folder, f'user_products/{product_id}/')
         if not os.path.exists(filepath):
             os.makedirs(filepath)
@@ -203,22 +234,18 @@ async def create_upload_file(product_id: int, file: UploadFile = File(...),
         if extension not in ["jpg", "png", "jpeg"]:
             return {"status": "error", "detail": "file extension not allowed"}
         token_name = secrets.token_hex(10) + "." + extension
-        generated_name = filepath + token_name
+        generated_name = os.path.join(filepath, token_name)
         file_content = await file.read()
-        with open(generated_name, "wb") as file:
-            file.write(file_content)
-        img = Image.open(generated_name)
-        img = img.resize(size=(200, 200))
-        img.save(generated_name)
-        file.close()
-        product = await UserProduct.get(id=product_id)
-        business = await product.business
-        if business.owner == user.id:
-            product.product_image = token_name
-            await product.save()
-        else:
-            raise UnauthorizedUserException
+        with open(generated_name, "wb") as f:
+            f.write(file_content)
+            img = Image.open(generated_name)
+            img = img.resize(size=(200, 200))
+            img.save(generated_name)
+            f.close()
+        product.product_image = token_name
+        await product.save()
         file_url = "localhost:8000" + generated_name[1:]
         return {"status": "ok", "filename": file_url}
     except Exception as e:
-        ServerErrorException(str(e))
+        raise ServerErrorException(str(e))
+
